test(runner): preserve stable test sharding
